reorganize taxonomy parsers and create common base
