+++ moonstone/parsers/counts/taxonomy/__init__.py
@@ -1,4 +1,3 @@
 from .kraken2 import SunbeamKraken2Parser  # noqa
-from .metaphlan2 import Metaphlan2Parser  # noqa
-from .metaphlan3 import Metaphlan3Parser  # noqa
+from .metaphlan import Metaphlan2Parser, Metaphlan3Parser  # noqa
 from .qiime import Qiime2Parser  # noqa
